Create PADUS_IUCN_SummaryBarCharts.R
This code creates a bar chart which summarizes CR, EN, and VU IUCN RL ecosystems that have at least 1% of their distribution falling inside the study areas. The charts show how much of the ecosystem overlaps with the study area and how protected it is inside and outside of the study area.